+++ src/views/normal/publicreport/index.vue
@@ -1,20 +1,8 @@
 <template>
   <div class="editor-container">
     <el-row :gutter="20">
-      <el-col :xs="24" :sm="24" :md="24" :lg="12">
+      <el-col :xs="24" :sm="24" :md="24" :lg="16">
         <div class="chart-wrapper">
-          <div class="verticalhalfdiv">
-            <div class="label">
-              <i class="el-icon-edit">
-                <a @click="openWindows('basic')" style="color: #66b1ff">
-                  場所名稱：</a
-                >
-              </i>
-            </div>
-            <div class="content">
-              <span> {{ this.buildinginfo.getName() }}</span>
-            </div>
-          </div>
           <div class="verticalhalfdiv">
             <div class="label">
               <i class="el-icon-edit">
@@ -25,59 +13,22 @@
             </div>
             <div class="content">
               <span class="report">
-                {{ TimeOptions("PublicSafeTimeOptions") }}
+                {{ TimeOptions("InspectionTimeOptions") }}
               </span>
             </div>
           </div>
         </div>
       </el-col>
-      <el-col :xs="24" :sm="24" :md="24" :lg="12">
-        <div class="chart-wrapper">
-          <el-col :xs="24" :sm="24" :md="24" :lg="12">
-            <div class="horizontalhalfdiv">
-              <div class="label">
-                <i class="el-icon-edit">
-                  <a @click="openWindows('user')" style="color: #66b1ff">
-                    管理權人：</a
-                  >
-                </i>
-              </div>
-              <div class="content">
-                <div
-                  v-for="(item, index) in this.buildinginfo.linkOwners"
-                  :key="index"
-                  class="user"
-                >
-                  <div style="padding-bottom: 2px">姓名 ： {{ item.name }}</div>
-                  <div>電話 ： {{ item.cellPhoneNumber }}</div>
-                </div>
-              </div>
-            </div>
-          </el-col>
-          <el-col :xs="24" :sm="24" :md="24" :lg="12">
-            <div class="horizontalhalfdiv">
-              <div class="label">
-                <i class="el-icon-edit">
-                  <a @click="openWindows('user')" style="color: #66b1ff">
-                    防火管理人：</a
-                  >
-                </i>
-              </div>
-              <div class="content">
-                <div
-                  v-for="(item, index) in this.buildinginfo.linkFireManagers"
-                  :key="index"
-                  class="user"
-                >
-                  <div style="padding-bottom: 2px">姓名 ： {{ item.name }}</div>
-                  <div>電話 ： {{ item.cellPhoneNumber }}</div>
-                </div>
-              </div>
-            </div>
-          </el-col>
+      <el-col :xs="24" :sm="24" :md="24" :lg="8">
+        <div class="right-wrapper">
+          <PanelGroup
+            :panelList="panelList"
+            v-on:handleSetLineChartData="handleSetLineChartData"
+          ></PanelGroup>
         </div>
       </el-col>
     </el-row>
+
     <el-row :gutter="32">
       <el-col :xs="24" :sm="24" :md="24" :lg="24">
         <div class="block-wrapper">
@@ -90,14 +41,6 @@
         </div>
       </el-col>
     </el-row>
-    <!-- <Dialog
-            v-if="innerVisible === true"
-            v-bind="dialogAttrs"
-            :files="files"
-            :formtableData="formtableData"
-            :formtableconfig="formtableconfig"
-            :listQueryParams="lacklistQueryParams"
-            v-on:handleDialog="handleDialog"></Dialog> -->
 
     <DialogForm
       ref="dialogform"
@@ -147,6 +90,14 @@ export default {
       //dialog額外的參數
       files: [],
       uploadVisible: false,
+      panelList: [],
+      lacksQueryParams: {
+        pageIndex: 1,
+        pageSize: 10,
+        total: 0,
+      },
+      lacksShow: false,
+      excludeId: null,
     };
   },
   computed: {
@@ -175,6 +126,7 @@ export default {
     async init() {
       this.title = "reportPublicSafe";
       this.tableConfig = PublicSafe.getTableConfig();
+      await this.getPublic();
       await this.getBuildingPublicSafeReport();
       await this.getProfessName();
       this.buttonsName = [
@@ -215,6 +167,53 @@ export default {
       this.dialogtableConfig = PublicSafeLack.getTableConfig();
       this.tableData = data.result;
       this.tablelistQueryParams.total = data.totalPageCount;
+    },
+    async getPublic() {
+      //取得公安申報 合格申報false & 改善狀況:未改善
+      var data = await PublicSafe.getSearchPage({
+        isImproved: false,
+        declareResult: false,
+        pageIndex: 1,
+        pageSize: 10,
+      });
+      if (this.setting_record == 0) {
+        await this.$store.dispatch("building/setoptions");
+        this.$store.dispatch("record/saveSettingRecord", 1);
+      }
+      let _array = this.buildingoptions.filter((item) => {
+        return (
+          item.classType == "LackStatusOptions" &&
+          item.textName == "已改善" &&
+          item.systemUse == true
+        );
+      });
+      console.log(_array);
+      var lacks = null;
+      if (_array.length) {
+        this.excludeId = _array[0].id;
+        this.lacksQueryParams.status = "{!=}" + _array[0].id;
+        lacks = await PublicSafeLack.getAllSearchPage(this.lacksQueryParams);
+      }
+      this.panelList = [
+        {
+          label: "缺失申報未改善",
+          count: data.totalPageCount,
+          svgIcon: "inspection",
+          type: "publicsafe",
+        },
+        {
+          label: "缺失項目未改善",
+          count: lacks !== null ? lacks.totalPageCount : 0,
+          svgIcon: "inspection",
+          type: "lacks",
+        },
+      ];
+    },
+    async getlacksSearchPage() {
+      this.lacksQueryParams.status = "{!=}" + this.excludeId;
+      var lack = await PublicSafeLack.getAllSearchPage(this.lacksQueryParams);
+      this.tableData = lack.result;
+      this.tablelistQueryParams.total = lack.totalPageCount;
     },
     async handleBlock(title, index, content) {
       //公安申報的操作
@@ -355,7 +354,20 @@ export default {
           this.$refs.block.clearSelectArray();
         }
       } else {
-        await this.handleLackDialog(title, index, content);
+        if (index === "updateLackManySave") {
+          var isOk = await content.update();
+          if (isOk) {
+            this.$message("更新成功");
+            await this.getPublicSafeLack();
+            this.dialogData.forEach((item, index) => {
+              if (item.id == content.id) {
+                this.dialogData.splice(index, 1, content);
+              }
+            });
+          }
+        } else {
+          await this.handleLackDialog(title, index, content);
+        }
       }
     },
     async handleLackDialog(title, index, content) {
@@ -368,9 +380,9 @@ export default {
             ? await content.create(this.publicSafe.getID())
             : await content.update();
         if (isOk) {
-          index === "updatelack"
-            ? this.$message("更新成功")
-            : this.$message("新增成功");
+          index === "createlack"
+            ? this.$message("新增成功")
+            : this.$message("更新成功");
           if (index === "createlack") this.isUpdate = true;
           await this.getPublicSafeLack();
           this.innerVisible = false;
@@ -396,9 +408,6 @@ export default {
           this.$message.error("系統錯誤");
         }
       }
-      // else if(index === 'selectData'){
-      //   this.$store.dispatch('building/setbuildingoptions',await Setting.getAllOption())
-      // }
     },
     async handleTableClick(index, content) {
       console.log(index, JSON.stringify(content));
@@ -413,8 +422,17 @@ export default {
         ];
         this.innerVisible = true;
         this.dialogStatus = "create";
-      } else if (index === "delete") {
-        var isDelete = await content.delete();
+      } else if (index === "delete" || index === "deleteMany") {
+        var isDelete = false;
+        if (index === "delete") {
+          isDelete = await content.delete();
+        } else {
+          var deleteArray = [];
+          content.forEach((item) => {
+            deleteArray.push(item.id);
+          });
+          isDelete = await PublicSafeLack.deleteMany(deleteArray.toString());
+        }
         if (isDelete) {
           this.$message("刪除成功");
           if (
@@ -425,7 +443,7 @@ export default {
               this.tablelistQueryParams.pageIndex - 1;
           }
           await this.getPublicSafeLack();
-          //await this.resettablelistQueryParams()
+          this.$refs.dialogtable.clearSelectArray();
           this.isUpdate = true;
         } else {
           this.$message.error("系統錯誤");
@@ -445,15 +463,51 @@ export default {
       } else if (index === "uploadExcel") {
         this.excelVisible = true;
         this.excelType = "uploadExcel";
+      } else if (index === "updateMany") {
+        this.dialogStatus = "updateMany";
+        content.forEach((item) => {
+          var obj = _.cloneDeep(item);
+          this.dialogData.push(obj);
+        });
+        this.dialogButtonsName = [
+          { name: "儲存", type: "primary", status: "updateLackManySave" },
+          { name: "取消", type: "info", status: "cancel" },
+        ];
+        this.innerVisible = true;
       } else if (index === "clickPagination") {
-        this.tablelistQueryParams = content;
-        await this.getPublicSafeLack();
+        if (this.lacksShow) {
+          this.lacksQueryParams = content;
+          await this.getlacksSearchPage();
+        } else {
+          this.tablelistQueryParams = content;
+          await this.getPublicSafeLack();
+        }
       } else {
         if (this.isUpdate) {
           await this.getBuildingPublicSafeReport();
           this.isUpdate = false;
         }
         this.tableVisible = false;
+        this.lacksShow = false;
+        this.lacksQueryParams = { pageIndex: 1, pageSize: 10, total: 0 };
+      }
+    },
+    async handleSetLineChartData(type) {
+      console.log("handleSetLineChartData", type);
+      if (type == "publicsafe") {
+        this.listQueryParams.isImproved = false;
+        this.listQueryParams.declareResult = false;
+        await this.getBuildingPublicSafeReport();
+      } else {
+        this.tableTitle = "publicsafelack";
+        this.dialogtableConfig = PublicSafeLack.getTableConfig();
+        this.tableheaderButtonsName = [
+          { name: "匯出檔案", icon: "el-icon-download", status: "exportExcel" },
+        ];
+        this.tablebuttonsName = [];
+        await this.getlacksSearchPage();
+        this.lacksShow = true;
+        this.tableVisible = true;
       }
     },
     async changeTable(value) {
@@ -468,9 +522,6 @@ export default {
 };
 </script>
 <style lang="scss" scoped>
-.el-form--inline .el-form-item {
-  margin-right: 40px;
-}
 .chart-wrapper {
   background: #fff;
   padding: 10px 10px;
@@ -479,54 +530,53 @@ export default {
   width: 100%;
   overflow-x: hidden;
   overflow-y: auto;
+
   .verticalhalfdiv {
     width: 100%;
     min-height: 50%;
     .label {
-      min-width: 30%;
+      min-width: 20%;
       display: inline-block;
       text-align: center;
       font-size: 20px;
-    }
-    .content {
-      min-width: 60%;
-      display: inline-block;
-      font-size: 30px;
-    }
-  }
-  .horizontalhalfdiv {
-    height: 100%;
-    min-width: 50%;
-    .label {
-      min-width: 30%;
-      display: inline-block;
-      text-align: center;
-      font-size: 20px;
-      vertical-align: top;
     }
     .content {
       min-width: 70%;
       display: inline-block;
-      font-size: 24px;
-      .user {
-        padding: 0px 8px 8px 8px;
-      }
+      font-size: 30px;
     }
   }
+
   .report {
     font-size: 50px;
     color: red;
   }
 }
-.chart-mwrapper {
-  background: #fff;
-  padding: 10px 16px 0;
-  margin-bottom: 32px;
-  height: 310px;
+
+.right-wrapper {
+  padding: 10px 10px;
+  margin-bottom: 30px;
+  height: 150px;
   width: 100%;
-  .report {
-    font-size: 24px;
-    color: red;
+  overflow-x: hidden;
+  overflow-y: auto;
+}
+
+@media (max-width: 1024px) {
+  .chart-wrapper {
+    padding: 8px;
+  }
+  .right-wrapper {
+    padding: 8px;
+  }
+  .left {
+    float: none;
+    width: 100%;
+  }
+  .right {
+    float: none;
+    width: 100%;
+    padding-left: 0px;
   }
 }
 </style>
